Surface Simplification 0.54
It's much faster now that I'm using hash sets instead of heaps, since
it's not very quick to update a heap.

Interface has changed so that you can try to specify a number of
triangles you want to end up with.

Triangles are being cleaned up, but vertices are not yet.